geoprocessing-worker/requirements.txt
# Core geoprocessing dependencies (from PoC)
rasterio>=1.3.0
Pillow>=9.0.0
numpy>=1.21.0
pyproj>=3.4.0
click>=8.0.0

# Worker-specific dependencies
httpx>=0.24.0
psycopg2-binary>=2.9.0
asyncio-mqtt>=0.13.0

# Logging and monitoring
structlog>=22.0.0

# Development and testing
pytest>=7.0.0
pytest-asyncio>=0.21.0